Yuval's Data added
signatures-1.txt - a list of information on  101 signatures, including
	k - the ephemeral key
	sc - the scalar used by openssl (this is k + once or twice the group
order)
	key - the secret key
	m - the hash of the message to sign (I randomly selected hashes)
	r,s - the signature
captured-1.txt - a list of 101 sequenced of doubles and adds captured
from signing the signatures above.  In these each line corresponds to a
signature.  D and A are
double and add and a dash (-) is a location where we may have missed
something